+++ OneBlink.SDK/models/FormPostSubmissionReceipt.cs
@@ -1,0 +1,10 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormPostSubmissionReceipt
+    {
+        public string html
+        {
+            get; set;
+        }
+    }
+}